rename package name queryMap to querymap
