+++ core/src/test/java/feign/querymap/PropertyQueryMapEncoderTest.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package feign.qeuryMap;
+package feign.querymap;
 
 import feign.QueryMapEncoder;
 import org.junit.Rule;
